backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
alembic
passlib[bcrypt]
pyjwt
python-multipart
pydantic[email]